requirements.txt
# Core AWS and SageMaker
sagemaker>=2.200.0
boto3>=1.28.0
botocore>=1.31.0

# ML Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0  # Alternative option

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0

# API and Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Monitoring and Logging
prometheus-client>=0.17.0
python-json-logger>=2.0.7
structlog>=23.1.0

# Data Validation and Configuration
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
moto>=4.1.0  # AWS mocking
httpx>=0.24.0  # FastAPI testing

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0
pylint>=2.17.0

# Utilities
tqdm>=4.66.0
click>=8.1.0
requests>=2.31.0
python-dateutil>=2.8.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Infrastructure as Code (optional)
aws-cdk-lib>=2.90.0
constructs>=10.0.0

# Security scanning
bandit>=1.7.5
safety>=2.3.0
